+++ app/routes/__init__.py
@@ -1,0 +1,5 @@
+"""
+Routing package consolidating all Flask blueprints under app.routes.*
+
+This structure keeps route modules cohesive and discoverable.
+"""
