+++ app/services/import_service.py
@@ -3,12 +3,14 @@ CSV/Excelインポート機能のサービスクラス
 """
 import pandas as pd
 import os
+import io
 from typing import Dict, List, Any, Optional
 from openpyxl import load_workbook
 from openpyxl.utils.exceptions import InvalidFileException
 from app import db
 from app.models import Project, Branch, ValidationError
 from app.enums import OrderProbability
+from flask import has_app_context
 
 
 class ImportService:
@@ -17,12 +19,12 @@ class ImportService:
     # 必須列の定義
     REQUIRED_COLUMNS = [
         'project_code',
-        'project_name', 
+        'project_name',
         'branch_name',
         'fiscal_year',
         'order_probability',
         'revenue',
-        'expenses'
+        'expenses',
     ]
     
     # 列名のマッピング（日本語 -> 英語）
@@ -30,33 +32,25 @@ class ImportService:
         'プロジェクトコード': 'project_code',
         'プロジェクト名': 'project_name',
         '支社名': 'branch_name',
-        '支社コード': 'branch_code',
         '売上の年度': 'fiscal_year',
-        '年度': 'fiscal_year',
         '受注角度': 'order_probability',
         '売上': 'revenue',
         '売上（契約金）': 'revenue',
         '契約金': 'revenue',
         '経費': 'expenses',
         '経費（トータル）': 'expenses',
-        'トータル経費': 'expenses'
     }
-    
-    # 受注角度の値マッピング
+
+    # 受注角度のマッピング（文字や記号 → 数値）
     ORDER_PROBABILITY_MAPPING = {
-        '〇': 100,
-        '○': 100,
+        '〇': 100, '○': 100, '◯': 100,
         '△': 50,
-        '×': 0,
-        'HIGH': 100,
-        'MEDIUM': 50,
-        'LOW': 0,
-        '100': 100,
-        '50': 50,
-        '0': 0,
-        100: 100,
-        50: 50,
-        0: 0
+        '×': 0, '✕': 0, '✖': 0,
+        'HIGH': 100, 'High': 100, 'high': 100,
+        'MEDIUM': 50, 'Medium': 50, 'medium': 50,
+        'LOW': 0, 'Low': 0, 'low': 0,
+        '100': 100, '50': 50, '0': 0,
+        100: 100, 50: 50, 0: 0,
     }
     
     def validate_file(self, filepath: str, file_type: str, sheet_name: str = None) -> Dict[str, Any]:
@@ -81,7 +75,7 @@ class ImportService:
                 df = pd.read_csv(filepath, encoding='utf-8-sig')
                 excel_info = None
             elif file_type == 'excel':
-                # Excelファイルの詳細情報を取得
+                # Excelファイルの詳細情報を取得（BytesIO経由でロック回避）
                 excel_info = self._get_excel_info(filepath)
                 if not excel_info['success']:
                     return excel_info
@@ -98,8 +92,11 @@ class ImportService:
                         'excel_info': excel_info
                     }
                 
-                # Excelファイルを読み込み
-                df = pd.read_excel(filepath, sheet_name=sheet_name)
+                # Excelファイルを読み込み（BytesIO + ExcelFile でファイルロック回避）
+                with open(filepath, 'rb') as f:
+                    data = f.read()
+                with pd.ExcelFile(io.BytesIO(data), engine='openpyxl') as xls:
+                    df = pd.read_excel(xls, sheet_name=sheet_name)
             else:
                 return {'success': False, 'error': 'サポートされていないファイル形式です'}
             
@@ -110,25 +107,20 @@ class ImportService:
             # 列名を正規化
             df.columns = df.columns.str.strip()
             
-            # 列名をマッピング
+            # 列名をマッピング（既知の日本語→英語は変換、未知は簡易正規化）
             mapped_columns = {}
             for col in df.columns:
                 if col in self.COLUMN_MAPPING:
                     mapped_columns[col] = self.COLUMN_MAPPING[col]
                 else:
                     mapped_columns[col] = col.lower().replace(' ', '_')
-            
             df = df.rename(columns=mapped_columns)
             
-            # 必須列の存在チェック
-            missing_columns = []
-            for required_col in self.REQUIRED_COLUMNS:
-                if required_col not in df.columns:
-                    missing_columns.append(required_col)
-            
-            if missing_columns:
+            # 必須列の存在チェック（CSV は列名が揃っていなくてもマッピング画面で対応するため緩和）
+            missing_columns = [c for c in self.REQUIRED_COLUMNS if c not in df.columns]
+            if file_type == 'excel' and missing_columns:
                 return {
-                    'success': False, 
+                    'success': False,
                     'error': f'必須列が不足しています: {", ".join(missing_columns)}',
                     'excel_info': excel_info
                 }
@@ -170,7 +162,10 @@ class ImportService:
             Dict: Excelファイル情報
         """
         try:
-            workbook = load_workbook(filepath, read_only=True, data_only=True)
+            # Windows のロック回避のため、一旦バイト列に読み込んでから openpyxl に渡す
+            with open(filepath, 'rb') as f:
+                data = f.read()
+            workbook = load_workbook(filename=io.BytesIO(data), read_only=True, data_only=True)
             
             sheets = []
             for sheet_name in workbook.sheetnames:
@@ -238,8 +233,11 @@ class ImportService:
                     excel_info = self._get_excel_info(filepath)
                     if excel_info['success'] and excel_info['sheets']:
                         sheet_name = excel_info['sheets'][0]['name']
-                
-                df = pd.read_excel(filepath, sheet_name=sheet_name)
+                # BytesIO + ExcelFile でクローズを保証しつつロック回避
+                with open(filepath, 'rb') as f:
+                    data = f.read()
+                with pd.ExcelFile(io.BytesIO(data), engine='openpyxl') as xls:
+                    df = pd.read_excel(xls, sheet_name=sheet_name)
             else:
                 return {'success': False, 'error': 'サポートされていないファイル形式です'}
             
@@ -254,6 +252,15 @@ class ImportService:
                     if file_column in df.columns:
                         mapped_df[system_field] = df[file_column]
                 df = mapped_df
+            else:
+                # 自動マッピング（日本語→英語、未知は簡易正規化）
+                auto_map = {}
+                for col in df.columns:
+                    if col in self.COLUMN_MAPPING:
+                        auto_map[col] = self.COLUMN_MAPPING[col]
+                    else:
+                        auto_map[col] = col.lower().replace(' ', '_')
+                df = df.rename(columns=auto_map)
             
             # データ検証と重複チェックを実行
             validation_result = self._validate_preview_data(df)
@@ -295,11 +302,16 @@ class ImportService:
         row_validations = {}
         duplicates = []
         
-        # 既存のプロジェクトコードを取得
+        # 既存のプロジェクトコードを取得（アプリコンテキストがない場合はスキップ）
         existing_codes = set()
-        existing_projects = Project.query.with_entities(Project.project_code).all()
-        for project in existing_projects:
-            existing_codes.add(project.project_code)
+        if has_app_context():
+            try:
+                existing_projects = Project.query.with_entities(Project.project_code).all()
+                for project in existing_projects:
+                    existing_codes.add(project.project_code)
+            except Exception:
+                # 取得失敗時は重複チェックをスキップ
+                existing_codes = set()
         
         # ファイル内の重複チェック
         if 'project_code' in df.columns:
@@ -411,8 +423,11 @@ class ImportService:
                     excel_info = self._get_excel_info(filepath)
                     if excel_info['success'] and excel_info['sheets']:
                         sheet_name = excel_info['sheets'][0]['name']
-                
-                df = pd.read_excel(filepath, sheet_name=sheet_name)
+                # BytesIO + ExcelFile でクローズを保証しつつロック回避
+                with open(filepath, 'rb') as f:
+                    data = f.read()
+                with pd.ExcelFile(io.BytesIO(data), engine='openpyxl') as xls:
+                    df = pd.read_excel(xls, sheet_name=sheet_name)
             else:
                 return {'success': False, 'error': 'サポートされていないファイル形式です'}
             
@@ -427,6 +442,15 @@ class ImportService:
                     if file_column in df.columns:
                         mapped_df[system_field] = df[file_column]
                 df = mapped_df
+            else:
+                # 自動マッピング（日本語→英語、未知は簡易正規化）
+                auto_map = {}
+                for col in df.columns:
+                    if col in self.COLUMN_MAPPING:
+                        auto_map[col] = self.COLUMN_MAPPING[col]
+                    else:
+                        auto_map[col] = col.lower().replace(' ', '_')
+                df = df.rename(columns=auto_map)
             
             # データ処理結果
             success_count = 0
@@ -936,21 +960,38 @@ class ImportService:
             Dict: 検証結果
         """
         try:
-            df = pd.read_excel(filepath, sheet_name=sheet_name)
+            # BytesIO + ExcelFile を使用してクローズを保証（ロック回避）
+            with open(filepath, 'rb') as f:
+                data = f.read()
+            with pd.ExcelFile(io.BytesIO(data), engine='openpyxl') as xls:
+                df = pd.read_excel(xls, sheet_name=sheet_name)
             
             if df.empty:
                 return {'success': False, 'error': f'シート「{sheet_name}」にデータが含まれていません'}
             
             # 列名を正規化
             df.columns = df.columns.str.strip()
+            original_columns = list(df.columns)
+            # 英語に正規化した列名も用意
+            mapped_columns = {}
+            for col in original_columns:
+                if col in self.COLUMN_MAPPING:
+                    mapped_columns[col] = self.COLUMN_MAPPING[col]
+                else:
+                    mapped_columns[col] = col.lower().replace(' ', '_')
+            normalized_df = df.rename(columns=mapped_columns)
             
             # サンプルデータを取得
-            sample_data = df.head(5).to_dict('records')
+            sample_data = normalized_df.head(5).to_dict('records')
+            
+            # 互換目的で columns には日本語と英語の両方を含める（順序は元→英語ユニオン）
+            union_columns = list(dict.fromkeys(original_columns + list(normalized_df.columns)))
             
             return {
                 'success': True,
-                'row_count': len(df),
-                'columns': list(df.columns),
+                'row_count': len(normalized_df),
+                'columns': union_columns,
+                'normalized_columns': list(normalized_df.columns),
                 'sample_data': sample_data
             }
