+++ scripts/init_db.py
@@ -3,6 +3,14 @@
 データベース初期化スクリプト
 """
 import os
+import sys
+from pathlib import Path
+
+# プロジェクトルートをパスに追加（scripts の親ディレクトリ）
+project_root = Path(__file__).resolve().parent.parent
+if str(project_root) not in sys.path:
+    sys.path.insert(0, str(project_root))
+
 from app import create_app, db
 from app.models import Project, Branch, FiscalYear
